requirements.txt
aiohttp==3.9.1
aiomysql==0.2.0
python-dotenv==1.0.0
tenacity==8.2.3
asyncio==3.4.3
certifi==2023.11.17